вынес метод convert() из MoveQuizViewController в Presenter
